Vars added, export added, logic for tests moved to separate files
