+++ set-vars.bat
@@ -1,7 +1,5 @@
-setx -m AUTH_VIEW_PHOTOS_FIELD "login"
-setx -m AUTH_ADD_POSTS_FIELD "login"
-setx -m AUTH_DELETE_POSTS_FIELD "login"
+setx -m AUTH_FIELD "login"
 
-setx -m AUTH_VIEW_PHOTOS_MATCH "gmail.com$"
-setx -m AUTH_ADD_POSTS_MATCH "gmail.com$"
-setx -m AUTH_DELETE_POSTS_MATCH "[help|main]@gmail.com"
+setx -m AUTH_VIEW_MATCH "gmail.com$"
+setx -m AUTH_ADD_MATCH "gmail.com$"
+setx -m AUTH_DELETE_MATCH "[help|main]@gmail.com"
