--- src/index.js
@@ -1,2 +1,0 @@
-//Only 'Pro' edition of Quokka.js supports project files import - how can modules be debugged?
-//TODO add main functionally here, e.g. routing
